+++ flink-learning-sql/flink-learning-sql-blink/src/main/java/com/zhisheng/sql/blink/stream/tableSource/KafkaSourceMain.java
@@ -1,8 +1,13 @@
-package com.zhisheng.sql.blink.stream.customTableSource;
+package com.zhisheng.sql.blink.stream.tableSource;
 
 import com.zhisheng.common.utils.ExecutionEnvUtil;
-import com.zhisheng.sql.blink.stream.customTableSink.MyRetractStreamTableSink;
+import com.zhisheng.common.utils.KafkaConfigUtil;
+import com.zhisheng.sql.blink.stream.tableSink.MyRetractStreamTableSink;
+import org.apache.flink.api.common.serialization.SimpleStringSchema;
+import org.apache.flink.api.java.utils.ParameterTool;
+import org.apache.flink.streaming.api.datastream.DataStream;
 import org.apache.flink.streaming.api.environment.StreamExecutionEnvironment;
+import org.apache.flink.streaming.connectors.kafka.FlinkKafkaConsumer011;
 import org.apache.flink.table.api.DataTypes;
 import org.apache.flink.table.api.EnvironmentSettings;
 import org.apache.flink.table.api.Table;
@@ -11,13 +16,15 @@ import org.apache.flink.table.sinks.RetractStreamTableSink;
 import org.apache.flink.table.types.DataType;
 import org.apache.flink.types.Row;
 
+import java.util.Properties;
+
 /**
- * Desc: blink custom kafka table source
- * Created by zhisheng on 2020-01-14 09:02
+ * Desc:
+ * Created by zhisheng on 2020-01-13 00:09
  * blog：http://www.54tianzhisheng.cn/
  * 微信公众号：zhisheng
  */
-public class CustomKafkaSourceMain {
+public class KafkaSourceMain {
     public static void main(String[] args) throws Exception {
         StreamExecutionEnvironment blinkStreamEnv = StreamExecutionEnvironment.getExecutionEnvironment();
         blinkStreamEnv.setParallelism(1);
@@ -27,7 +34,11 @@ public class CustomKafkaSourceMain {
                 .build();
         StreamTableEnvironment blinkStreamTableEnv = StreamTableEnvironment.create(blinkStreamEnv, blinkStreamSettings);
 
-        blinkStreamTableEnv.registerTableSource("kafkaDataStream", new MyKafkaTableSource(ExecutionEnvUtil.PARAMETER_TOOL));
+        ParameterTool parameterTool = ExecutionEnvUtil.PARAMETER_TOOL;
+        Properties properties = KafkaConfigUtil.buildKafkaProps(parameterTool);
+        DataStream<String> dataStream = blinkStreamEnv.addSource(new FlinkKafkaConsumer011<>(parameterTool.get("kafka.topic"), new SimpleStringSchema(), properties));
+        Table table = blinkStreamTableEnv.fromDataStream(dataStream, "word");
+        blinkStreamTableEnv.registerTable("kafkaDataStream", table);
 
         RetractStreamTableSink<Row> retractStreamTableSink = new MyRetractStreamTableSink(new String[]{"_count", "word"}, new DataType[]{DataTypes.BIGINT(), DataTypes.STRING()});
         blinkStreamTableEnv.registerTableSink("sinkTable", retractStreamTableSink);
@@ -36,6 +47,6 @@ public class CustomKafkaSourceMain {
 
         wordCount.insertInto("sinkTable");
 
-        blinkStreamTableEnv.execute("Blink Custom Kafka Table Source");
+        blinkStreamTableEnv.execute("Blink Kafka Table Source");
     }
 }
