+++ flink-learning-sql/flink-learning-sql-blink/src/main/java/com/zhisheng/sql/blink/stream/tableSource/MyKafkaTableSource.java
@@ -1,4 +1,4 @@
-package com.zhisheng.sql.blink.stream.customTableSource;
+package com.zhisheng.sql.blink.stream.tableSource;
 
 import com.zhisheng.common.utils.KafkaConfigUtil;
 import org.apache.flink.api.common.serialization.SimpleStringSchema;
